fix: UserAddress 위경도 타입 불일치 해결 (BigDecimal -> Double)
- DB의 float8 타입과 매칭되도록 엔티티 및 DTO의 `lat`, `lng` 타입을 `Double`로 변경
- Schema-validation 에러 해결을 위해 불필요한 Flyway 마이그레이션 파일 삭제 및 정리

+++ src/main/java/com/mygomi/backend/api/dto/request/AddressRequestDto.java
@@ -24,8 +24,8 @@ public class AddressRequestDto {
     private String banchi;
 
     @Schema(description = "위도")
-    private BigDecimal lat;
+    private Double lat;
 
     @Schema(description = "경도")
-    private BigDecimal lng;
+    private Double lng;
 }
